chore: class changes for TimeTree
